subscription_monitor/subscriptions/requirements.txt
Django==4.2.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
matplotlib==3.8.0
pandas==2.1.3
numpy==1.24.3
beautifulsoup4==4.12.2
html2text==2020.1.16